Don't ignore tar failures
When the user is `root` (as is often the case in a container), then
`tar` will try to preserve the ownership and permissions of the files
since it's operating in backup restoration mode. But these are just
source files and you'd actually prefer to ignore the mode in the tarball
as would be done for an unprivileged user. Explicitly pass
`--no-same-owner --no-same-permissions` as is the default for
unprivileged users. This is what `dpkg` and other packaging tools do
when extracting source tarballs.